support rerun of rscripts on opencpu. Each package needs to have a saveData function to save the data on opencpu server (see storeParameterToOpencpuServer function).  The session objects of the saved data are saved for rerun. A guid is created for each rscript run. The guid as well as the id of the tab of the page are stored and passed to the gaggle_output.html page.
Once user initiates a rerun from the gaggle_output.html page, the guid and tab id together with new parameter values are sent to background page, which passes it to gaggle.js of the source page using the tab id. We fill the parameters and rerun the rscript on opencpu. Results are passed to the gaggle_output page.